feat(exceptions.py): add MongOrmException base class and three subclasses: DocumentNotFound, DuplicateDocument, and MissingClientException
refactor(mongorm.py): remove unused imports and move exceptions to separate module
